Fixed random crashes when zoning. Was due to improper write of the SessionEndPacket. For extra safety db is locked until zone in happens. Added some logs. Seems to be all working!

+++ FFXIVClassic World Server/Packets/WorldPackets/Send/SessionEndPacket.cs
@@ -45,7 +45,7 @@ namespace FFXIVClassic_World_Server.Packets.WorldPackets.Send
                     try
                     {
                         binWriter.Write((UInt32)destinationZoneId);
-                        binWriter.Write((UInt32)spawnType);
+                        binWriter.Write((UInt16)spawnType);
                         binWriter.Write((Single)spawnX);
                         binWriter.Write((Single)spawnY);
                         binWriter.Write((Single)spawnZ);
